Initial version
This is at the point where we can get to the individual Objects within
the Library file. Decoding the name table is half done. I've converted
the C example hash routing (not fun as it used pointers) to compute the
string hashes, so I will be able to validate the name table data.